--- src/access.json
@@ -1,5 +1,0 @@
-{
-	"symfony": "",
-	"tvstRemember": "",
-	"user": 0
-}
