+++ test/index.js
@@ -1,35 +1,35 @@
-// const tvnode = require('../src/index')
-// const utils = require('../src/utils')
+const tvnode = require('../src/index')
+const utils = require('../src/utils')
 
-// utils.setCookie({ 'symfony': '', 'tvstRemember': '' })
+utils.setCookie({ 'symfony': '', 'tvstRemember': '' })
 
-// utils.setUser()
+utils.setUser()
 
-// console.info(utils.getCookies())
+console.info(utils.getCookies())
 
-// console.info(utils.getUser())
+console.info(utils.getUser())
 
-// tvnode.login('hola', '123', true)
+tvnode.login('hola', '123', true)
 
-// tvnode.shows()
-//   .then(d => {
-//     console.info(d)
-//   })
+tvnode.shows()
+  .then(d => {
+    console.info(d)
+  })
 
-// tvnode.show(260449)
-//   .then(d => {
-//     console.info(d.seasons[0])
-//   })
-//   .catch(err => {
-//     console.error(err)
-//   })
+tvnode.show(260449)
+  .then(d => {
+    console.info(d.seasons[0])
+  })
+  .catch(err => {
+    console.error(err)
+  })
 
-// tvnode.episode(260449, 4383248)
-//   .then(d => {
-//     console.info(d)
-//   })
+tvnode.episode(260449, 4383248)
+  .then(d => {
+    console.info(d)
+  })
 
-// tvnode.episodeWatch(4383246)
-//   .then(d => {
-//     console.info(d)
-//   })
+tvnode.episodeWatch(4383246)
+  .then(d => {
+    console.info(d)
+  })
