+++ tsconfig.json
@@ -1,0 +1,22 @@
+{
+    "compilerOptions": {
+      "target": "es6",
+      "lib": [
+        "ESNext",
+        "DOM"
+      ],                         
+      "module": "commonjs",
+      "declaration": true,                     
+      "declarationMap": true,                
+      "sourceMap": true,             
+      "outDir": "./dist",                        
+      "resolveJsonModule": true,
+      "strict": true,                           
+      "esModuleInterop": true,                  
+      "skipLibCheck": true,                     
+      "forceConsistentCasingInFileNames": true  
+    },
+    "include": [
+      "src/**/*"
+    ]
+  }
